@@ -1,13 +1,13 @@
 name: simple_mapper
 description: Maps objects of different types to other objects using configured expressions.
-version: 1.1.0
+version: 1.2.0
 repository: https://github.com/jacobmcgowan/simple_mapper
 issue_tracker: https://github.com/jacobmcgowan/simple_mapper/issues
 homepage: https://github.com/jacobmcgowan/simple_mapper
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.0.0
-  test: ^1.0.0
+  pedantic: ^1.11.0
+  test: ^1.17.7
